# Conflicts:
#	GolemEditor/Include/UI/Windows/viewport.h
#	GolemEditor/Source/UI/Windows/viewport.cpp
#	GolemEngine/Include/Viewport/scene.h
#	GolemEngine/Source/Viewport/scene.cpp
#	GolemEngine/Source/inputManager.cpp

+++ GolemEditor/Source/UI/Windows/fileBrowser.cpp
@@ -12,9 +12,12 @@
 #include "ImGuiFileDialog-master/ImGuiFileDialog.h"
 #include "Wrappers/windowWrapper.h"
 #include "Resource/tools.h"
+#include "Viewport/scene.h"
+#include "Core/gameobject.h"
 
 namespace fs = std::filesystem;
 
+// Avoid deleting important files that users do not have access to
 #define EXCLUDE_FILE(name) \
     (name != "x64" && \
     name != "GolemEditor.vcxproj" && \
@@ -46,6 +49,9 @@ void FileBrowser::Update(GolemEngine* _golemEngine)
 	ImGui::EndChild();
 	RightMouseClickEvent();
 	ImGui::End();
+	
+	// Drag and drop event
+	DragandDropEvent();
 }
 
 
@@ -81,6 +87,7 @@ void FileBrowser::TreeNodes(std::filesystem::path _path)
 
 void FileBrowser::ContentBrowser()
 {
+	// Make it on sameline
 	ImGui::Text(GetFileName(m_currentDirectory.string().c_str()));
 	ImGui::Text("");
 	if (m_currentDirectory != m_editorDirectory)
@@ -96,53 +103,114 @@ void FileBrowser::ContentBrowser()
 		ImGui::Text("");
 	}
 
+	if (!isLoadUi)
+	{
+		// Load every file's default UI
+		Ui_Default = WindowWrapper::LoadUiTexture(Tools::FindFile("default_Ui.png").c_str());
+		Ui_Folder = WindowWrapper::LoadUiTexture(Tools::FindFile("File_Icon.png").c_str());
+		Ui_Cpp = WindowWrapper::LoadUiTexture(Tools::FindFile("cpp_Icon.png").c_str());;
+		Ui_H = WindowWrapper::LoadUiTexture(Tools::FindFile("h_Icon.png").c_str());
+		Ui_Obj = WindowWrapper::LoadUiTexture(Tools::FindFile("obj_Icon.png").c_str());
+		// Do once
+		isLoadUi = true;
+	}
+
+	// For displaying all files in this folder path
 	for (auto& p : fs::directory_iterator(m_currentDirectory))
 	{
 		std::string path = p.path().string();
 		std::string fileName = GetFileName(path.c_str());
+		// Get the extension like .obj .cpp .h ....
 		std::string extensionFile = GetFileExtension(fileName);
 		if (EXCLUDE_FILE(fileName))
 		{
+			// Every file is a small imgui window
 			ImGui::BeginChild(GetFileName(path.c_str()), ImVec2(100, 100));
-			ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.0f, 0.0f, 0.0f, 0.0f));
-			if (ImGui::Button(GetFileName(path.c_str()), ImVec2(70, 70)))
+			// Check the mouse is on the UI or not
+			// if it is on the UI show the button 
+			if (ImGui::IsMouseHoveringRect(ImGui::GetWindowPos(), ImVec2(ImGui::GetWindowPos().x + ImGui::GetWindowSize().x, ImGui::GetWindowPos().y + ImGui::GetWindowSize().y)))
 			{
-				if (p.is_directory())
+				ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.0f, 0.0f, 0.0f, 0.0f));
+				// the UI is a button that we can click it to tigger new event
+				if (ImGui::Button(GetFileName(path.c_str()), ImVec2(70, 70)))
 				{
-					m_currentDirectory = path;
+					double currentTime = ImGui::GetTime();
+					double clickInterval = currentTime - m_buttonState.lastClickTime;
+					m_buttonState.lastClickTime = currentTime;
+					if (p.is_directory() && clickInterval < ImGui::GetIO().MouseDoubleClickTime)
+					{
+						m_currentDirectory = path;
+						m_buttonState.clicked = false;
+					}
 				}
+				ImGui::PopStyleColor();
+				ImGui::SetCursorPosY(ImGui::GetCursorPosY() - 70);
 			}
-			ImGui::PopStyleColor();
-			ImGui::SetCursorPosY(ImGui::GetCursorPosY() - 70);
+			// Right click to open content menu
 			if (ImGui::IsMouseReleased(ImGuiMouseButton_Right) && ImGui::IsItemHovered())
 			{
-				std::cout << "right clicked this window" << path.c_str() << std::endl;
-				// TODO a small menu
 				ImGui::OpenPopup("FolderContextMenu");
-				SelectedFolder = path;
+				selectedFolder = path;
 			}
 
+			if (ImGui::BeginDragDropSource())
+			{
+				draggingFilePath = path;
+				ImGui::SetDragDropPayload("FileDrop", draggingFilePath.c_str(), strlen(draggingFilePath.c_str()) + 1);
+				isDragging = true;
+				ImGui::EndDragDropSource();
+			}
+
+			// Show UI ICON
 			if (p.is_directory())
 			{
-				Golemint texture = WindowWrapper::LoadUiTexture(Tools::FindFile("File_Icon.png").c_str());
-				ImGui::Image((void*)(intptr_t)texture, ImVec2(70, 70));
+				ImGui::Image((void*)(intptr_t)Ui_Folder, ImVec2(70, 70));
 			}
+			// Show Texture image
 			else if (extensionFile == ".jpg" || extensionFile == ".png")
 			{
-				Golemint texture = WindowWrapper::LoadUiTexture(path.c_str());
-				ImGui::Image((void*)(intptr_t)texture, ImVec2(70, 70));
+				auto it = std::find_if(m_loadedTextures.begin(), m_loadedTextures.end(), [&](const std::pair<std::string, GLuint>& element) {
+					return element.first == path;
+					});
+				if (it != m_loadedTextures.end()) {
+					ImGui::Image((void*)(intptr_t)it->second, ImVec2(70, 70));
+				}
+				else {
+					GLuint texture = WindowWrapper::LoadUiTexture(path.c_str());
+					ImGui::Image((void*)(intptr_t)texture, ImVec2(70, 70));
+					m_loadedTextures.push_back(std::make_pair(path, texture)); 
+				}
+			}
+			else if (extensionFile == ".cpp")
+			{
+				ImGui::Image((void*)(intptr_t)Ui_Cpp, ImVec2(70, 70));
+			}
+			else if (extensionFile == ".h")
+			{
+				ImGui::Image((void*)(intptr_t)Ui_H, ImVec2(70, 70));
+			}
+			else if (extensionFile == ".obj")
+			{
+				ImGui::Image((void*)(intptr_t)Ui_Obj, ImVec2(70, 70));
 			}
 			else
 			{
-				Golemint texture = WindowWrapper::LoadUiTexture(Tools::FindFile("File_Icon.png").c_str());
-				ImGui::Image((void*)(intptr_t)texture, ImVec2(70, 70));
+				ImGui::Image((void*)(intptr_t)Ui_Default, ImVec2(70, 70));
 			}
+			// Show content menu
+			// Menu selections:
+			// Rename
+			// Delete
 			ImGui::Text(GetFileName(path.c_str()));
 			if (ImGui::BeginPopupContextItem("FolderContextMenu"))
 			{
+				if (ImGui::MenuItem("Rename"))
+				{
+					// TODO
+				}
 				if (ImGui::MenuItem("Delete"))
 				{
-					DeleteFolder(SelectedFolder.c_str());
+					DeleteFolder(selectedFolder.c_str());
 				}
 				ImGui::EndPopup();
 			}
@@ -152,6 +220,7 @@ void FileBrowser::ContentBrowser()
 	}
 }
 
+// Record the previous file path for the back button
 void FileBrowser::LastPath(std::filesystem::path& _currentPath)
 {
 	if (!_currentPath.empty())
@@ -164,7 +233,7 @@ void FileBrowser::LastPath(std::filesystem::path& _currentPath)
 		_currentPath = m_editorDirectory;
 	}
 }
-
+// Get the file name with after the last "\"
 const char* FileBrowser::GetFileName(const char* _path)
 {
 	if (_path == nullptr)
@@ -181,7 +250,7 @@ const char* FileBrowser::GetFileName(const char* _path)
 	return _path + index + 1;
 }
 
-
+// Get the file extentension like .obj .cpp .h ......
 std::string FileBrowser::GetFileExtension(const std::string& _fileName) 
 {
 	size_t dotPosition = _fileName.find_last_of('.');
@@ -191,6 +260,7 @@ std::string FileBrowser::GetFileExtension(const std::string& _fileName)
 	return "";
 }
 
+// Right click in file viewer
 void FileBrowser::RightMouseClickEvent()
 {
 	if (ImGui::IsItemHovered() && ImGui::IsMouseClicked(1))
@@ -208,12 +278,15 @@ void FileBrowser::RightMouseClickEvent()
 	}
 }
 
+// Create a new folder 
 void FileBrowser::CreateFolder()
 {
+	// with default name "NewFolder"
 	std::string newFolderName = "NewFolder";
 	
 	int count = 1;
 
+	// If it already exsits "NewFolder", add a number at the tail
 	while (std::filesystem::exists(m_currentDirectory / newFolderName))
 	{
 		newFolderName = "NewFolder" + std::to_string(count);
@@ -234,11 +307,45 @@ void FileBrowser::DeleteFolder(const std::string& _folderPath)
 {
 	try
 	{
-		std::filesystem::remove_all(_folderPath);
-		std::cout << "Folder " << _folderPath << "deleted successfully." << std::endl;
+		const char* folderName = GetFileName(_folderPath.c_str());
+		// Avoid deleting the important folders
+		const std::vector<const char*> protectedFolders = { "Assets", "Source", "Include", "Shaders" };
+
+		if (std::find(protectedFolders.begin(), protectedFolders.end(), folderName) == protectedFolders.end())
+		{
+			std::filesystem::remove_all(_folderPath);
+			std::cout << "Folder " << _folderPath << " deleted successfully." << folderName << std::endl;
+		}
 	}
-	catch(const std::exception& e)
+	catch (const std::exception& e)
 	{
 		std::cerr << "Failed to delete folder: " << e.what() << std::endl;
+	}
+
+}
+
+void FileBrowser::LoadFile(const std::string& _filePath)
+{
+	// TODO
+}
+
+void FileBrowser::DragandDropEvent()
+{
+	if (isDragging)
+	{
+		ImVec2 mousePos = ImGui::GetMousePos();
+		ImVec2 windowPos = ImVec2(mousePos.x - 35, mousePos.y - 35);
+
+		ImGui::SetNextWindowPos(windowPos);
+		ImGui::Begin("Dragging Window", nullptr, ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoTitleBar);
+		// Show a drop Icon UI
+		ImGui::Image((void*)(intptr_t)Ui_Default, ImVec2(70, 70));
+		ImGui::End();
+	}
+	// If we released the mouse left button will trigger event
+	if (ImGui::IsMouseReleased(0))
+	{
+		isDragging = false;
+		std::cout << GetFileName(draggingFilePath.c_str()) << std::endl;
 	}
 }
